# session,executor: performance improvement in SetProcessInfo