--- a/session/session.go
+++ b/session/session.go
@@ -865,7 +865,7 @@ func (s *session) SetProcessInfo(sql string, t time.Time, command byte) {
 	pi := util.ProcessInfo{
 		ID:      s.sessionVars.ConnectionID,
 		DB:      s.sessionVars.CurrentDB,
-		Command: mysql.Command2Str[command],
+		Command: string(command),
 		Time:    t,
 		State:   s.Status(),
 		Info:    sql,
@@ -1628,6 +1628,8 @@ func (s *session) ShowProcess() util.ProcessInfo {
 	tmp := s.processInfo.Load()
 	if tmp != nil {
 		pi = tmp.(util.ProcessInfo)
+		byteCommand := []byte(pi.Command)
+		pi.Command = mysql.Command2Str[byteCommand[0]]
 	}
 	return pi
 }