--- a/util/processinfo.go
+++ b/util/processinfo.go
@@ -23,7 +23,7 @@ type ProcessInfo struct {
 	User    string
 	Host    string
 	DB      string
-	Command string
+	Command byte
 	Time    time.Time
 	State   uint16
 	Info    string