@ngaut 2019-03-12
This is not a good design, could you change the type of Command from string to byte, and delay the conversion from byte to string when we actually need to.